requirements.txt
streamlit>=1.38.0
google-genai>=1.0.0
elevenlabs>=1.10.0
plotly>=5.20.0
pandas>=2.1.0
python-dotenv>=1.0.0
